# specs/002-web-dashboard/openapi.yaml
openapi: 3.0.3
info:
  title: Lukas the Bear Admin Dashboard API
  description: |
    REST API for the Lukas the Bear Admin Dashboard. Provides monitoring and control
    capabilities for the Slack chatbot including activity logs, generated images,
    scheduled events, and manual control triggers.

    ## Authentication
    All endpoints (except login) require a valid session cookie obtained through
    the POST /api/auth/login endpoint. Sessions remain valid until the browser is closed.

    ## Rate Limiting
    Manual control endpoints are rate-limited to prevent abuse (max 10 requests per hour per session).
  version: 1.0.0
  contact:
    name: Admin Dashboard Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://dashboard.lukas-bear.example.com
    description: Production server
...
paths:
...
  /api/images:
    get:
      tags:
        - Generated Images
      summary: Get paginated image gallery
      description: |
        Returns a paginated list of AI-generated images with thumbnails and metadata.
        Results are ordered by creation timestamp descending (most recent first).
      operationId: getImages
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: start_date
          in: query
          description: Filter images from this date onwards (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-10-01T00:00:00Z"
        - name: end_date
          in: query
          description: Filter images until this date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-10-28T23:59:59Z"
        - name: status
          in: query
          description: Filter by image status
          schema:
            type: string
            enum:
              - generated
              - posted
              - failed
          example: "posted"
      responses:
        '200':
          description: Image gallery retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageSummary'
                  total:
                    type: integer
                    description: Total number of matching items
                    example: 42
                  page:
                    type: integer
                    description: Current page number
                    example: 1
                  pages:
                    type: integer
                    description: Total number of pages
                    example: 3
              examples:
                success:
                  value:
                    items:
                      - id: "img-789"
                        created_at: "2025-10-28T12:00:00Z"
                        prompt: "A friendly bear enjoying autumn leaves"
                        status: "posted"
                        posted_to_channel: "C01234567"
                        posted_at: "2025-10-28T12:01:30Z"
                        thumbnail_url: "/api/images/img-789/thumbnail"
                      - id: "img-790"
                        created_at: "2025-10-27T18:30:00Z"
                        prompt: "Celebration with confetti and balloons"
                        status: "failed"
                        posted_to_channel: null
                        posted_at: null
                        thumbnail_url: "/api/images/img-790/thumbnail"
                    total: 42
                    page: 1
                    pages: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/images/{id}:
    get:
      tags:
        - Generated Images
      summary: Get full image details
      description: |
        Returns complete details for a specific generated image including full resolution URL,
        generation parameters, cost information, and error messages for failed attempts.
      operationId: getImageDetail
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Image ID
          schema:
            type: string
          example: "img-789"
      responses:
        '200':
          description: Image details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDetail'
              examples:
                success:
                  value:
                    id: "img-789"
                    created_at: "2025-10-28T12:00:00Z"
                    prompt: "A friendly bear enjoying autumn leaves in a cozy forest setting"
                    image_url: "file:///app/data/images/img-789.png"
                    status: "posted"
                    posted_to_channel: "C01234567"
                    posted_at: "2025-10-28T12:01:30Z"
                    generation_duration_seconds: 8.3
                    cost_usd: 0.04
                    error_message: null
                    metadata:
                      theme: "seasonal_autumn"
                      model: "dall-e-3"
                      size: "1024x1024"
                      quality: "standard"
                      slack_message_ts: "1698494490.123456"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/images/{id}/thumbnail:
    get:
      tags:
        - Generated Images
      summary: Get image thumbnail
      description: |
        Returns a JPEG thumbnail (300x300px) of the generated image. Thumbnails are
        generated on-demand and cached to the filesystem for subsequent requests.
      operationId: getImageThumbnail
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Image ID
          schema:
            type: string
          example: "img-789"
      responses:
        '200':
          description: Thumbnail image
          headers:
            Cache-Control:
              description: Caching directive
              schema:
                type: string
                example: "max-age=86400"
            Content-Type:
              description: Image MIME type
              schema:
                type: string
                example: "image/jpeg"
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  # Scheduled Events Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie obtained from POST /api/auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Short error identifier
          example: "Unauthorized"
        message:
          type: string
          description: Human-readable error message
          example: "Authentication required"
      required:
        - error
        - message
...
    ImageSummary:
      type: object
      description: Summary of a generated image for gallery listing
      properties:
        id:
          type: string
          description: Unique image ID
          example: "img-789"
        created_at:
          type: string
          format: date-time
          description: When the image was generated (ISO 8601)
          example: "2025-10-28T12:00:00Z"
        prompt:
          type: string
          description: Prompt used for image generation
          example: "A friendly bear enjoying autumn leaves"
        status:
          type: string
          enum:
            - generated
            - posted
            - failed
          description: Current status of the image
          example: "posted"
        posted_to_channel:
          type: string
          nullable: true
          description: Slack channel ID where image was posted (null if not posted)
          example: "C01234567"
        posted_at:
          type: string
          format: date-time
          nullable: true
          description: When the image was posted (ISO 8601, null if not posted)
          example: "2025-10-28T12:01:30Z"
        thumbnail_url:
          type: string
          description: URL to fetch thumbnail (300x300 JPEG)
          example: "/api/images/img-789/thumbnail"
      required:
        - id
        - created_at
        - prompt
        - status
        - thumbnail_url
    ImageDetail:
      type: object
      description: Full details of a generated image
      properties:
        id:
          type: string
          description: Unique image ID
          example: "img-789"
        created_at:
          type: string
          format: date-time
          description: When the image was generated (ISO 8601)
          example: "2025-10-28T12:00:00Z"
        prompt:
          type: string
          description: Full prompt used for image generation
          example: "A friendly bear enjoying autumn leaves in a cozy forest setting"
        image_url:
          type: string
          description: Full resolution image URL (file:// or http://)
          example: "file:///app/data/images/img-789.png"
        status:
          type: string
          enum:
            - generated
            - posted
            - failed
          description: Current status of the image
          example: "posted"
        posted_to_channel:
          type: string
          nullable: true
          description: Slack channel ID where image was posted
          example: "C01234567"
        posted_at:
          type: string
          format: date-time
          nullable: true
          description: When the image was posted (ISO 8601)
          example: "2025-10-28T12:01:30Z"
        generation_duration_seconds:
          type: number
          format: float
          nullable: true
          description: Time taken to generate the image in seconds
          example: 8.3
        cost_usd:
          type: number
          format: float
          nullable: true
          description: Cost of image generation in USD
          example: 0.04
        error_message:
          type: string
          nullable: true
          description: Error message if generation or posting failed
          example: null
        metadata:
          type: object
          nullable: true
          description: Additional metadata (theme, model, size, quality, slack message timestamp)
          example:
            theme: "seasonal_autumn"
            model: "dall-e-3"
            size: "1024x1024"
            quality: "standard"
            slack_message_ts: "1698494490.123456"
      required:
        - id
        - created_at
        - prompt
        - image_url
        - status
...
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
                message: "Authentication required. Please log in."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                error: "Not found"
                message: "The requested resource does not exist"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            serverError:
              value:
                error: "Internal server error"
                message: "An unexpected error occurred. Please try again later."